+++ src/django_routines/exceptions.py
@@ -1,0 +1,4 @@
+class ExitEarly(Exception):
+    """
+    Raise this exception to exit a routine early.
+    """