+++ tests/django_routines_tests/management/commands/keyerr.py
@@ -1,0 +1,7 @@
+import typing as t
+from django.core.management import BaseCommand
+
+
+class Command(BaseCommand):
+    def handle(self, *args, **options) -> t.Optional[str]:
+        raise KeyError()